pyproject.toml
[project]
name = "vibe-coding"
version = "0.1.0"
description = "Automated trading system with Blue Short Strategy and Fubon Securities API integration"
authors = [
    {name = "Yixiu1121",email = "a0986811099@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "schedule>=1.2.0",
    "loguru>=0.7.0",
    "pytz>=2023.3",
    "psycopg2-binary>=2.9.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "python-dotenv>=1.0.0",
    "fubon-neo @ file:///Users/yixiu/Downloads/fubon_neo-2.2.4-cp37-abi3-macosx_11_0_arm64.whl"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vibe-coding"
version = "0.1.0"
description = "Automated trading system with Blue Short Strategy and Fubon Securities API integration"
authors = ["Yixiu1121 <a0986811099@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
numpy = "^1.24.0"
pyyaml = "^6.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
schedule = "^1.2.0"
loguru = "^0.7.0"
pytz = "^2023.3"
psycopg2-binary = "^2.9.0"
matplotlib = "^3.5.0"
seaborn = "^0.11.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
